requirements.txt
absl-py==2.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cycler==0.12.1
# -e git+https://github.com/f1tenth/f1tenth_gym.git@4fdb9c7e6fb7c701290f4dc18377d07c1681724f#egg=f110_gym
Farama-Notifications==0.0.4
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
future==1.0.0
grpcio==1.76.0
gym==0.21.0
importlib-metadata==4.13.0
Jinja2==3.1.6
kiwisolver==1.4.9
llvmlite==0.45.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
pyglet==1.5.20
PyOpenGL==3.1.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scipy==1.15.3
six==1.17.0
stable-baselines3==1.8.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.9.1
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
